+++ NEWS.md
@@ -1,3 +1,12 @@
+# whichWithin 0.7.0
+
+* `which_within_dist_dura` now handles unique
+
+
+# whichWithin 0.6.0
+
+* New function: `which_within_dist_dura` for distance/duration scans
+
 # whichWithin 0.5.0
 
 * Now in C